Refactoring tests to increase execution speed.
