+++ src/test/java/com/scmspain/infrastructure/controller/PublishTweetCommandITCase.java
@@ -1,0 +1,62 @@
+package com.scmspain.infrastructure.controller;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
+import org.springframework.web.context.WebApplicationContext;
+
+import com.scmspain.infrastructure.configuration.TestConfiguration;
+
+import static java.lang.String.format;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = TestConfiguration.class)
+public class PublishTweetCommandITCase {
+
+    @Autowired
+    private WebApplicationContext context;
+
+    private MockMvc mockMvc;
+
+    @Before
+    public void setUp() {
+        this.mockMvc = webAppContextSetup(this.context).build();
+    }
+
+    @Test
+    public void shouldReturn200WhenInsertingAValidTweet() throws Exception {
+        mockMvc
+            .perform(newTweet("Prospect", "Breaking the law"))
+            .andExpect(status().is(201));
+    }
+
+    @Test
+    public void shouldReturn200WhenInsertingAValidTweetWithUrls() throws Exception {
+        mockMvc
+            .perform(newTweet("Prospect", "Breaking the law: http://www.judaspriest.com http://www.judaspriest.com http://www.judaspriest.com http://www.judaspriest.com http://www.judaspriest.com http://www.judaspriest.com"))
+            .andExpect(status().is(201));
+    }
+
+    @Test
+    public void shouldReturn400WhenInsertingAnInvalidTweet() throws Exception {
+        mockMvc
+            .perform(newTweet("Schibsted Spain", "We are Schibsted Spain (look at our home page http://www.schibsted.es/), we own Vibbo, InfoJobs, fotocasa, coches.net and milanuncios. Welcome! Text added to make it fail."))
+            .andExpect(status().is(400));
+    }
+
+    private MockHttpServletRequestBuilder newTweet(String publisher, String tweet) {
+        return post("/tweet")
+                .contentType(MediaType.APPLICATION_JSON_UTF8)
+                .content(format("{\"publisher\": \"%s\", \"tweet\": \"%s\"}", publisher, tweet));
+    }
+
+}
